+++ src/interfaces/mocks/mod.rs
@@ -1,0 +1,7 @@
+//! # Mocks
+//!
+//! Mock Interfaces used for testing of coordinator library functionality
+
+pub mod clientchain;
+pub mod service;
+pub mod storage;
