+++ src/interfaces/mod.rs
@@ -8,3 +8,6 @@ pub mod request;
 pub mod response;
 pub mod service;
 pub mod storage;
+
+#[cfg(test)]
+pub mod mocks;
